feat(Sparkles): move Gradient to utils
